Creating Views

+++ Balance/Balance/Models/CreditModel.cs
@@ -7,7 +7,7 @@ namespace Balance.Models
 {
     public class CreditModel
     {
-        public int Money { get; set; }
+        public Decimal Money { get; set; }
         //public int DebtorId { get; set; }
         //public int CreditorId { get; set; }
     }
