+++ Balance/Entities/Credit.cs
@@ -9,7 +9,7 @@ namespace Entities
     public class Credit
     {
         public Guid Id { get; set; }
-        public int Money { get; set; }
+        public Decimal Money { get; set; }
         //public int DebtorId { get; set; }
         //public int CreditorId { get; set; }
     }
